lrdp-v2: Fix a use-before-init of mem_section
With the latest page tracking script, mem_section will not be
initialized to a valid value in function parse(), unless Linux
kernel version is greater than 3.19. This makes the script
reporting an error when page owner info is parsed for msm-3.18,
and thus page tracking info not saved successfully.

Assign the default value to mem_section unconditionally before
kernel version is checked.

Change-Id: I60961ab204e0e4f4aac7213de3a80a5ed091e4a6